Recreate a web project with Brunch support
