+++ apps/web/.gitignore
@@ -7,6 +7,14 @@
 # Generated on crash by the VM
 erl_crash.dump
 
+# Static artifacts
+/node_modules
+
+# Since we are building assets from web/static,
+# we ignore priv/static. You may want to comment
+# this depending on your deployment strategy.
+/priv/static/
+
 # The config/prod.secret.exs file by default contains sensitive
 # data and you should not commit it into version control.
 #
