--- apps/web/web/controllers/map_controller.ex
@@ -1,8 +1,0 @@
-defmodule Web.MapController do
-  use Web.Web, :controller
-
-  def index(conn, _params) do
-    render(conn, "index.html")
-  end
-
-end
