--- apps/web/web/css/styles.css
@@ -1,47 +1,0 @@
-body {
-  background-color: #b5d0d0;
-  margin: 0;
-  padding: 0;
-  font-family: 'Roboto', sans-serif;
-}
-
-#map {
-  height: 100vh;
-  /* width: 100vW; */
-}
-
-.awesome-marker i {
-    font-size: 16px;
-    margin-top: 8px;
-}
-
-.left-panel {
-    width: 200px;
-    height: 100vh;
-    float: left;
-    border-right: solid 2px #b29ac3;
-}
-
-.simulator {
-}
-
-.simulator label {
-    width: 100%;
-}
-
-.simulator input {
-    width: 90%;
-}
-
-.simulator button {
-    width: 48%;
-}
-
-.geolocation {
-    margin-top: 1em;
-}
-
-.geolocation button {
-    width: 100%;
-}
-
