--- apps/web/web/index.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <title>Escubidu</title>
-
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-
-    <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
-    <link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
-    <link rel="stylesheet" href="css/ionicons.min.css">
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.0.2/dist/leaflet.css" />
-    <link rel="stylesheet" href="css/leaflet.awesome-markers.css">
-    <link rel="stylesheet" href="css/styles.css" />
-
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-</head>
-
-<body>
-    <div class="left-panel">
-        <div class="simulator">
-            <form>
-                <label for="latitude">Latitude:</label>
-                <input type="text" id="latitude" name="latitude" placeholder="40.4169"/>
-
-                <label for="longitude">Longitude:</label>
-                <input type="text" id="longitude" name="longitude" placeholder="-3.7035"/>
-
-                <button id="send">Send</button>
-                <button id="render">Render</button>
-            </form>
-        </div>
-
-        <div class="geolocation">
-            <form>
-                <button id="geolocate">Geolocate yourself</button>
-            </form>
-            <ul id="locations"/>
-        </div>
-    </div>
-
-    <div id="map"></div>
-
-    <script src="./js/jquery-3.1.1.min.js"></script>
-    <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
-    <script src="https://unpkg.com/leaflet@1.0.2/dist/leaflet.js"></script>
-    <script src="./js/leaflet.awesome-markers.js"></script>
-    <script src="./js/gpx.js"></script>
-    <script src="./js/leaflet-hash.js"></script>
-    <script src="./js/routes.js"></script>
-    <script src="./js/app.js"></script>
-    <script src="./js/simulator.js"></script>
-    <script src="./js/geolocation.js"></script>
-</body>
-
-</html>
