--- apps/web/web/js/routes.js
@@ -1,14 +1,0 @@
-var routes = [
-    {
-        "type": "En bici",
-        "icon": "android-bicycle",
-        "color": "green",
-        "gpx": [ ]
-    },
-    {
-        "type": "En coche",
-        "icon": "android-car",
-        "color": "red",
-        "gpx": [ ]
-    }
-];
