+++ apps/web/web/router.ex
@@ -17,7 +17,6 @@ defmodule Web.Router do
     pipe_through :browser # Use the default browser stack
 
     get "/", PageController, :index
-    get "/map", MapController, :index
   end
 
   # Other scopes may use custom stacks.
