+++ apps/web/web/static/css/app.css
@@ -1,0 +1,1 @@
+/* This file is for your main application css. */
