+++ apps/web/web/static/css/phoenix.css
@@ -80,6 +80,3 @@ form.link, form.button {
     border-bottom: 0;
   }
 }
-/* This file is for your main application css. */
-
-/*# sourceMappingURL=app.css.map*/
