--- apps/web/web/templates/map/index.html.eex
@@ -1,2 +1,0 @@
-<h2>The map will be shown here</h2>
-
