--- apps/web/web/views/map_view.ex
@@ -1,3 +1,0 @@
-defmodule Web.MapView do
-  use Web.Web, :view
-end
